LayoutTests/imported/w3c/web-platform-tests/html/syntax/serializing-html-fragments/template-expected.txt


PASS Template element content is correctly serialized
FAIL HTML fragment serialization algorithm should be applied to the template content assert_equals: expected "<div xml:lang=\"en-us\" p:attr=\"v\"></div>" but got "<div xmlns=\"xx\" xml:lang=\"en-us\" p:attr=\"v\" xmlns:p=\"uri2\"/>"
